fix(runtime): make Host process bridge build on Windows
Move process-group Kill/Start into unix/windows files so release
cross-compiles for windows/amd64 and windows/arm64. Harden Runtime
Activity sandbox tests to wait for running status under CI load.
